Added conf/db.yml to store db config options, including a reload Redis on statrtup option
Added a delete function to Redis shortcake

Added group relationship to Pid

Solidified testing for the users and pid controllers

Tweaked UI for user management screens

+++ app.rb
@@ -4,9 +4,12 @@
 $LOAD_PATH.unshift(File.absolute_path(File.join(File.dirname(__FILE__), 'lib/shortcake')))
 require 'shortcake'
 
+config = YAML.load_file('conf/db.yml')
+
 class PidApp < Sinatra::Application
   enable :sessions # enable cookie-based sessions
   set :session_secret, 'super secret'
+  set :sessions, :expire_after => 900
   
   set :root, File.dirname(__FILE__)
   
@@ -19,8 +22,10 @@ class PidApp < Sinatra::Application
     ENV['DATABASE_URL'] ||= "sqlite3://#{File.absolute_path(File.dirname(__FILE__))}/db/dev.db"
   end
   
-  configure :seeded do
-    ENV['DATABASE_URL'] ||= "sqlite3://#{File.absolute_path(File.dirname(__FILE__))}/db/seeded.db"
+  configure :seeded do  
+#    ENV['DATABASE_URL'] ||= "sqlite3://#{File.absolute_path(File.dirname(__FILE__))}/db/seeded.db"
+    ENV['DATABASE_URL'] = "mysql://root:@localhost/seeded"
+    
   end
   
   configure :test do
@@ -53,9 +58,11 @@ class PidApp < Sinatra::Application
 end
 
 # set database
+puts "Establishing connection to the database"
 DataMapper.setup(:default, ENV['DATABASE_URL'])
 
 # load controllers and models
+puts "Building controllers and models" 
 Dir.glob("controllers/*.rb").each { |r| require_relative r }
 Dir.glob("models/*.rb").each { |r| require_relative r }
 
@@ -70,6 +77,15 @@ end
 
 # OPTIMIZE - Should this go here?
 #reload the Redis database from the data stored in the DB
-shorty = Shortcake.new('pid', {:host => 'localhost', :port => 6379})
-shorty.flushall!
-Pid.all.each { |pid| shorty.create(pid.id.to_s, pid.url) if !pid.deactivated }
+if config['rebuild_redis_on_startup']
+  puts "Rebuilding the Redis database for pid resolution"
+  shorty = Shortcake.new('pid', {:host => 'localhost', :port => 6379})
+  shorty.flushall!
+  Pid.all.each do |pid| 
+    begin
+      shorty.create(pid.id.to_s, pid.url) if !pid.deactivated 
+    rescue Exception => e
+      puts "something happened while rebuilding the Redis DB for PID #{pid.id}: #{e.message}"
+    end
+  end
+end
